fix tests

+++ ewm-service/src/main/java/ru/practicum/explore/ewm/events/dto/LocationDto.java
@@ -5,7 +5,6 @@ import lombok.*;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -16,9 +15,9 @@ public class LocationDto {
     @Max(value = 90)
     @Min(value = -90)
     @NotNull
-    private BigDecimal lat;
+    private Double lat;
     @Max(value = 180)
     @Min(value = -180)
     @NotNull
-    private BigDecimal lon;
+    private Double lon;
 }
